Refactor `cargo-binstall`
 - Move everything in `src/main.rs` to the lib to speedup building
 - Move `logging` to `binstalk` to reduce codegen for `cargo-binstall`
