+++ crates/bin/Cargo.toml
@@ -31,20 +31,14 @@ file-format = { version = "0.18.0", default-features = false }
 fs-lock = { version = "0.1.0", path = "../fs-lock" }
 gh-token = "0.1.2"
 home = "0.5.5"
-log = { version = "0.4.18", features = ["std"] }
 miette = "5.9.0"
 mimalloc = { version = "0.1.37", default-features = false, optional = true }
-once_cell = "1.18.0"
 semver = "1.0.17"
 strum = "0.25.0"
 strum_macros = "0.25.0"
-supports-color = "2.0.0"
 tempfile = "3.5.0"
 tokio = { version = "1.28.2", features = ["rt-multi-thread", "signal"], default-features = false }
-tracing-core = "0.1.31"
 tracing = { version = "0.1.37", default-features = false }
-tracing-log = { version = "0.1.3", default-features = false }
-tracing-subscriber = { version = "0.3.17", features = ["fmt", "json", "ansi"], default-features = false }
 
 [build-dependencies]
 embed-resource = "2.1.1"
@@ -78,11 +72,11 @@ cross-lang-fat-lto = ["binstalk/cross-lang-fat-lto"]
 fancy-no-backtrace = ["miette/fancy-no-backtrace"]
 fancy-with-backtrace = ["fancy-no-backtrace", "miette/fancy"]
 
-log_max_level_info = ["log/max_level_info", "tracing/max_level_info", "log_release_max_level_info"]
-log_max_level_debug = ["log/max_level_debug", "tracing/max_level_debug", "log_release_max_level_debug"]
+log_max_level_info = ["binstalk/log_max_level_info"]
+log_max_level_debug = ["binstalk/log_max_level_debug"]
 
-log_release_max_level_info = ["log/release_max_level_info", "tracing/release_max_level_info"]
-log_release_max_level_debug = ["log/release_max_level_debug", "tracing/release_max_level_debug"]
+log_release_max_level_info = ["binstalk/log_release_max_level_info"]
+log_release_max_level_debug = ["binstalk/log_release_max_level_debug"]
 
 [package.metadata.docs.rs]
 rustdoc-args = ["--cfg", "docsrs"]
