+++ crates/bin/src/lib.rs
@@ -1,11 +1,13 @@
 #![cfg_attr(docsrs, feature(doc_auto_cfg))]
 
-pub mod args;
-pub mod bin_util;
-pub mod entry;
+mod args;
+mod bin_util;
+mod entry;
 mod git_credentials;
 mod install_path;
-pub mod logging;
 mod manifests;
 mod signal;
 mod ui;
+
+mod do_main;
+pub use do_main::main_impl;
