+++ crates/binstalk/Cargo.toml
@@ -23,15 +23,21 @@ either = "1.8.1"
 itertools = "0.11.0"
 jobslot = { version = "0.2.11", features = ["tokio"] }
 leon = { version = "2.0.1", path = "../leon" }
+log = { version = "0.4.18", features = ["std"] }
 maybe-owned = "0.3.4"
 miette = "5.9.0"
+once_cell = "1.18.0"
 semver = { version = "1.0.17", features = ["serde"] }
 strum = "0.25.0"
+supports-color = "2.0.0"
 target-lexicon = { version = "0.12.11", features = ["std"] }
 tempfile = "3.5.0"
 thiserror = "1.0.40"
 tokio = { version = "1.30.0", features = ["rt", "process", "sync"], default-features = false }
 tracing = "0.1.37"
+tracing-core = "0.1.31"
+tracing-log = { version = "0.1.3", default-features = false }
+tracing-subscriber = { version = "0.3.17", features = ["fmt", "json", "ansi"], default-features = false }
 url = { version = "2.3.1", features = ["serde"] }
 
 [features]
@@ -56,6 +62,12 @@ http3 = ["binstalk-downloader/http3"]
 
 zstd-thin = ["binstalk-downloader/zstd-thin"]
 cross-lang-fat-lto = ["binstalk-downloader/cross-lang-fat-lto"]
+
+log_max_level_info = ["log/max_level_info", "tracing/max_level_info", "log_release_max_level_info"]
+log_max_level_debug = ["log/max_level_debug", "tracing/max_level_debug", "log_release_max_level_debug"]
+
+log_release_max_level_info = ["log/release_max_level_info", "tracing/release_max_level_info"]
+log_release_max_level_debug = ["log/release_max_level_debug", "tracing/release_max_level_debug"]
 
 [package.metadata.docs.rs]
 rustdoc-args = ["--cfg", "docsrs"]
